+++ R/image.R
@@ -19,11 +19,13 @@ image_to_bytes <- function(im) {
   res        <- list()
   res$width  <- ncol(im)
   res$height <- nrow(im)
+  res$realalpha <- FALSE
   
   if (is.matrix(im)) {
     res$colorspace <- '/DeviceGray' 
     res$pixels     <- t(im)
     res$alpha      <- rep(255, res$width * res$height) # Fake alpha
+    res$realalpha  <- FALSE
   } else if (is.array(im)) {
     stopifnot(length(dim(im)) == 3)
     nplanes <- dim(im)[3]
@@ -32,21 +34,25 @@ image_to_bytes <- function(im) {
       res$colorspace <- '/DeviceGray' 
       res$pixels     <- t(im)
       res$alpha      <- rep(255, res$width * res$height) # Fake alpha
+      res$realalpha  <- FALSE
     } else if (nplanes == 2) {
       res$colorspace <-  '/DeviceGray'
       res$pixels <- im[,,1] |> t()
       res$alpha  <- im[,,2] |> t()
+      res$realalpha <- TRUE
     } else if (nplanes == 3) {
       res$pixels     <- aperm(im, c(3, 2, 1)) |> as.vector()
       res$colorspace <- '/DeviceRGB'
       res$alpha      <- rep(255, res$width * res$height) # Fake alpha
+      res$realalpha  <- FALSE
     } else if (nplanes == 4) {
       res$colorspace <- '/DeviceRGB'
       alpha          <- im[, , 4]
       im             <- im[, , 1:3]
       res$pixels     <- aperm(im, c(3, 2, 1)) |> as.vector()
       res$colorspace <- '/DeviceRGB'
-      res$alpha       <- t(alpha)
+      res$alpha      <- t(alpha)
+      res$realalpha  <- TRUE
     } else {
       stop("Unhandled num of planes: ", nplanes)
     }