+++ NEWS.md
@@ -1,12 +1,15 @@
 
 
-# minipdf 0.2.5.9001 2025-08-11
+# minipdf 0.2.5.9002 2025-08-12
 
 * Replace object-oriented R6 approach with pipe-able approach
 * Add support for
     * multi-page documents
     * per-text font settings
     * global and local transforms
+    * global and local clipping
+    * images
+    * Filters: Ascii85, Hex, RLE
 
 * Introduction of pipe-able PDF document creation.